WIP: Refactoring - removed useless daos - renamed some folders
