--- quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/ContextType.java
@@ -1,6 +1,0 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql;
-
-public enum ContextType {
-    WORKSPACE,
-    DATASET
-}
