+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sparql/SPARQLContextType.java
@@ -1,0 +1,6 @@
+package fr.cnrs.liris.jpugetgil.sparqltosql.sparql;
+
+public enum SPARQLContextType {
+    WORKSPACE,
+    DATASET
+}
