+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/SQLVarType.java
@@ -1,4 +1,4 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql;
 
 public enum SQLVarType {
     DATA,
