+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/comparison/ComparisonOperator.java
@@ -1,4 +1,4 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.comparison;
 
 public interface ComparisonOperator {
     String getComparisonOperatorName();
