+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/comparison/EqualToOperator.java
@@ -1,4 +1,4 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.comparison;
 
 public class EqualToOperator implements ComparisonOperator {
     @Override
