+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/comparison/LessThanOperator.java
@@ -1,4 +1,4 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.comparison;
 
 public class LessThanOperator implements ComparisonOperator {
     @Override
