+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/comparison/NotEqualToOperator.java
@@ -1,4 +1,4 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.comparison;
 
 public class NotEqualToOperator implements ComparisonOperator {
     @Override
