+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/operator/StSBGPOperator.java
@@ -1,10 +1,11 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.context;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
 
 import com.github.jsonldjava.shaded.com.google.common.collect.Streams;
-import fr.cnrs.liris.jpugetgil.sparqltosql.*;
-import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLClause;
-import fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator.EqualToOperator;
-import fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator.NotEqualToOperator;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sparql.SPARQLOccurrence;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sparql.SPARQLPositionType;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.*;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.comparison.EqualToOperator;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.comparison.NotEqualToOperator;
 import org.apache.jena.graph.*;
 import org.apache.jena.sparql.algebra.op.OpBGP;
