+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/operator/StSGraphOperator.java
@@ -1,9 +1,9 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.context;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
 
-import fr.cnrs.liris.jpugetgil.sparqltosql.SQLContext;
-import fr.cnrs.liris.jpugetgil.sparqltosql.SQLQuery;
-import fr.cnrs.liris.jpugetgil.sparqltosql.SQLVarType;
-import fr.cnrs.liris.jpugetgil.sparqltosql.SQLVariable;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLContext;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLQuery;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLVarType;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLVariable;
 import org.apache.jena.graph.Node_Variable;
 import org.apache.jena.sparql.algebra.op.OpGraph;
