+++ quads-query/src/main/java/fr/cnrs/liris/jpugetgil/sparqltosql/sql/operator/StSOperator.java
@@ -1,7 +1,7 @@
-package fr.cnrs.liris.jpugetgil.sparqltosql.sql.context;
+package fr.cnrs.liris.jpugetgil.sparqltosql.sql.operator;
 
-import fr.cnrs.liris.jpugetgil.sparqltosql.SQLQuery;
-import fr.cnrs.liris.jpugetgil.sparqltosql.SQLVariable;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLQuery;
+import fr.cnrs.liris.jpugetgil.sparqltosql.sql.SQLVariable;
 
 import java.util.ArrayList;
 import java.util.List;
